client/requirements.txt
httpx>=0.25.0
soundfile>=0.12.1
numpy>=1.26.0
PyAudioWPatch>=0.2.12
python-dotenv>=1.0.1
